custom select flexible for array of strings / objects; update to public-profile-stup/services page
