+++ assets/styles/auth-screens.module.css
@@ -239,7 +239,6 @@
 
 .form_group .custom_country_select .countries {
   padding: var(--sm-inner-component-mg) 0;
-  /* border: 1px solid var(--grey-4); */
   border-radius: var(--input-br);
   box-shadow: 0px 1px 3px 0px rgba(50, 50, 71, 0.1),
     0px 0px 1px 0px rgba(12, 26, 75, 0.2);
@@ -254,7 +253,6 @@
   background-color: #fff;
 }
 
-/* hide countries container scroll bar */
 .form_group .custom_country_select .countries::-webkit-scrollbar {
   width: 0 !important;
 }
@@ -266,7 +264,6 @@
 .form_group .custom_country_select .countries {
   -ms-overflow-style: none;
 }
-/* end */
 
 .form_group .custom_country_select .countries .input_wrapper {
   position: sticky;
@@ -276,10 +273,8 @@
 }
 
 .form_group .custom_country_select .countries input[type="search"] {
-  /* height: 40px; */
   margin: 0 var(--sm-inner-component-mg);
   width: calc(100% - var(--sm-inner-component-mg) * 2);
-  /* calc to get inner br */
   border-radius: calc(var(--input-br) - var(--sm-inner-component-mg));
 }
 
@@ -344,7 +339,6 @@
 .form_group.error textarea,
 .form_group.error select {
   border: 1px solid var(--input-error-border) !important;
-  /* box-shadow: 0px 0px 0px 3px var(--input-error-box-shadow); */
   background-color: var(--input-error-bg) !important;
 }
 
@@ -355,6 +349,31 @@
   border: 1px solid var(--input-error-border) !important;
   background-color: var(--input-error-bg) !important;
 }
+
+.form_group select {
+  appearance: none;
+  -moz-appearance: none;
+  -webkit-appearance: none;
+  background: #fff
+    url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='30' height='30' viewBox='0 0 20 20' fill='none'%3E%3Cpath d='M16.6 7.45837L11.1667 12.8917C10.525 13.5334 9.47499 13.5334 8.83333 12.8917L3.39999 7.45837' stroke='rgb(189, 192, 206)' strokeWidth='2.5' strokeMiterlimit='10' strokeLinecap='round' strokeLinejoin='round'/%3E%3C/svg%3E")
+    no-repeat right var(--input-padding) center/1rem !important;
+}
+
+.form_group select:focus {
+  appearance: none;
+  -moz-appearance: none;
+  -webkit-appearance: none;
+  background: #fff
+    url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='30' height='30' viewBox='0 0 20 20' fill='none'%3E%3Cpath d='M16.6 7.45837L11.1667 12.8917C10.525 13.5334 9.47499 13.5334 8.83333 12.8917L3.39999 7.45837' stroke='rgb(28, 17, 50)' strokeWidth='2.5' strokeMiterlimit='10' strokeLinecap='round' strokeLinejoin='round'/%3E%3C/svg%3E")
+    no-repeat right var(--input-padding) center/1rem !important;
+}
+
+.form_group.error select {
+  background: var(--input-error-bg)
+    url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='30' height='30' viewBox='0 0 20 20' fill='none'%3E%3Cpath d='M16.6 7.45837L11.1667 12.8917C10.525 13.5334 9.47499 13.5334 8.83333 12.8917L3.39999 7.45837' stroke='rgb(189, 192, 206)' strokeWidth='2.5' strokeMiterlimit='10' strokeLinecap='round' strokeLinejoin='round'/%3E%3C/svg%3E")
+    no-repeat right var(--input-padding) center/1rem !important;
+}
+
 
 .form .action_ctn {
   margin: 48px 0 var(--inner-component-mg);
@@ -396,29 +415,6 @@
   background-color: var(--input-focus-bg);
 }
 
-.form_group select {
-  appearance: none;
-  -moz-appearance: none;
-  -webkit-appearance: none;
-  background: #fff
-    url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='30' height='30' viewBox='0 0 20 20' fill='none'%3E%3Cpath d='M16.6 7.45837L11.1667 12.8917C10.525 13.5334 9.47499 13.5334 8.83333 12.8917L3.39999 7.45837' stroke='rgb(189, 192, 206)' strokeWidth='2.5' strokeMiterlimit='10' strokeLinecap='round' strokeLinejoin='round'/%3E%3C/svg%3E")
-    no-repeat right var(--input-padding) center/1rem !important;
-}
-
-.form_group select:focus {
-  appearance: none;
-  -moz-appearance: none;
-  -webkit-appearance: none;
-  background: #fff
-    url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='30' height='30' viewBox='0 0 20 20' fill='none'%3E%3Cpath d='M16.6 7.45837L11.1667 12.8917C10.525 13.5334 9.47499 13.5334 8.83333 12.8917L3.39999 7.45837' stroke='rgb(28, 17, 50)' strokeWidth='2.5' strokeMiterlimit='10' strokeLinecap='round' strokeLinejoin='round'/%3E%3C/svg%3E")
-    no-repeat right var(--input-padding) center/1rem !important;
-}
-
-.form_group.error select {
-  background: var(--input-error-bg)
-    url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='30' height='30' viewBox='0 0 20 20' fill='none'%3E%3Cpath d='M16.6 7.45837L11.1667 12.8917C10.525 13.5334 9.47499 13.5334 8.83333 12.8917L3.39999 7.45837' stroke='rgb(189, 192, 206)' strokeWidth='2.5' strokeMiterlimit='10' strokeLinecap='round' strokeLinejoin='round'/%3E%3C/svg%3E")
-    no-repeat right var(--input-padding) center/1rem !important;
-}
 
 /* input icons */
 .input_wrapper {
