+++ components/Auth/SignupLayoutLHS/signup-layout-lhs.module.css
@@ -125,10 +125,51 @@
   }
 }
 
+@media screen and (max-width: 1000px) {
+  .auth_level.mobile_width_1000 {
+    min-height: unset;
+    gap: 5px
+  }
+
+  .auth_content_lhs .inner {
+    display: flex;
+    width: 100%;
+    flex-direction: row;
+    /* justify-content: space-between; */
+    justify-content: center;
+  }
+
+  .auth_level.mobile_width_1000 {
+    display: flex;
+    /* flex: 1 1; */
+  }
+
+  .auth_level.mobile_width_1000 .lhs {
+    flex-direction: row;
+    position: relative;
+  }
+
+  .auth_level.mobile_width_1000 .lhs .round-ctn {
+    z-index: 4;
+  }
+
+  .auth_level.mobile_width_1000 .lhs .line {
+    width: 50px;
+    /* position: absolute; */
+    left: calc(100% - 30px);
+    height: 3px;
+    z-index: 3;
+  }
+
+  .auth_level.mobile_width_1000 .content {
+    display: none;
+  }
+}
+
 @media screen and (max-width: 700px) {
   .auth_level {
     min-height: unset;
-    gap: 5px
+    gap: 5px;
   }
 
   .auth_content_lhs {
