+++ components/FileUpload/index.module.css
@@ -3,11 +3,19 @@
   border: 1px solid var(--grey-5);
   border-radius: var(--input-br);
   padding: var(--inner-component-mg);
-  margin-bottom: var(--inner-component-mg);
   box-shadow: var(--input-box-shadow);
   display: flex;
   flex-direction: column;
   gap: var(--outer-component-mg);
+}
+
+.upload_ctn.error {
+  border: 1px solid var(--input-error-border) !important;
+}
+
+.upload_ctn.error .file_upload {
+  border: 1px solid var(--input-error-border) !important;
+  background-color: var(--input-error-bg) !important;
 }
 
 .upload_ctn .file_upload,
